requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
requests==2.31.0
python-jose[cryptography]==3.3.0
einops==0.7.0
tqdm==4.66.1
accelerate==0.24.1
transformers==4.35.2
diffusers==0.21.4
pydantic==2.5.0
opencv-python-headless==4.8.1.78
# TryOnDiffusion dependencies (code is included locally)
# git+https://github.com/fashn-AI/tryondiffusion.git